Fix for ticket creation process

+++ src/main/java/com/bics/jira/mail/IssueLookupHelper.java
@@ -11,7 +11,7 @@ import com.atlassian.jira.service.util.handler.MessageHandlerErrorCollector;
  * @since 04.02.13 21:57
  */
 public interface IssueLookupHelper {
-    MutableIssue lookupByKey(String text, int resolvedBefore, MessageHandlerErrorCollector monitor);
+    MutableIssue lookupByKey(String text, long resolvedBefore, MessageHandlerErrorCollector monitor);
 
-    MutableIssue lookupBySubject(Project project, String subject, int resolvedBefore, MessageHandlerErrorCollector monitor);
+    MutableIssue lookupBySubject(Project project, String subject, long resolvedBefore, MessageHandlerErrorCollector monitor);
 }
